moved searchSaga and searchSlice to the Search component in the common folder
