+++ src/features/people/People/index.js
@@ -2,7 +2,7 @@ import { useSelector } from "react-redux";
 import Loading from "../../../common/Navigation/Search/Loading";
 import Error from "../../../common/Navigation/Search/Error";
 import { PopularPeopleList } from "../PopularPeopleList";
-import { selectQuery, selectSearchResult, selectSearchStatus } from "../searchSlice";
+import { selectQuery, selectSearchResult, selectSearchStatus } from "../../../common/Navigation/Search/searchSlice";
 import NoResult from "../../../common/Navigation/Search/NoResult";
 
 export const People = () => {
